fix: fix forbidden error without headers
