+++ .gitignore
@@ -1,3 +1,5 @@
+downloads/
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
