+++ .vscode/settings.json
@@ -1,0 +1,7 @@
+{
+    "files.associations": {
+        "__split_buffer": "cpp",
+        "deque": "cpp",
+        "vector": "cpp"
+    }
+}
